Make PhoneProvider a SimpleProvider
Services have a cost in memory, and there was no point running PhoneProvider as a service since it doesn't have to store any state.
