+++ app/src/main/AndroidManifest.xml
@@ -147,7 +147,6 @@
 
         <service android:name=".dataprovider.AppProvider" />
         <service android:name=".dataprovider.ContactsProvider" />
-        <service android:name=".dataprovider.PhoneProvider" />
         <service android:name=".dataprovider.SearchProvider" />
         <service android:name=".dataprovider.SettingsProvider" />
         <service android:name=".dataprovider.ShortcutsProvider" />
