+++ app/src/main/res/xml/preferences.xml
@@ -227,10 +227,6 @@
             android:title="@string/settings_name" />
         <fr.neamar.kiss.SwitchPreference
             android:defaultValue="true"
-            android:key="enable-phone"
-            android:title="@string/phone_name" />
-        <fr.neamar.kiss.SwitchPreference
-            android:defaultValue="true"
             android:key="enable-shortcuts"
             android:title="@string/shortcuts_name" />
         <fr.neamar.kiss.SwitchPreference
